+++ logprep/util/rule_dry_runner.py
@@ -1,24 +1,63 @@
 #!/usr/bin/python3
-"""This module runs the pipeline for specified events and shows how processing changed them."""
+"""
+Dry Run
+-------
+
+Rules can be tested by executing them in a dry run of Logprep.
+Instead of the connectors defined in the configuration file,
+the dry run takes a path parameter to an input JSON (line) file that contains log messages.
+The output is displayed in the console and changes made by Logprep are being highlighted:
+
+..  code-block:: bash
+    :caption: Directly with Python
+
+    PYTHONPATH="." python3 logprep/run_logprep.py $CONFIG --dry-run $EVENTS
+
+..  code-block:: bash
+    :caption: With a PEX file
+
+    logprep.pex $CONFIG --dry-run $EVENTS
+
+Where :code:`$CONFIG` is the path to a configuration file
+(see :doc:`configuration/configurationdata`).
+The only required section in the configuration is :code:`pipeline`
+(see tests/testdata/config/config-dry-run.yml for an example).
+The remaining options are set internally or are being ignored.
+
+:code:`$EVENTS` is the path to a file with one or multiple log messages.
+A single log message can be provided with a file containing a plain json or wrapped in brackets
+(beginning with `[` and ending with `]`).
+For multiple events it must be a list wrapped inside brackets, while each log object separated by a
+comma.
+By specifying the parameter :code:`--dry-run-input-type jsonl` a list of JSON lines can be used
+instead.
+Additional output, like pseudonyms, will be printed if :code:`--dry-run-full-output` is added.
+
+..  code-block:: bash
+    :caption: Example for execution with a JSON lines file (dry-run-input-type jsonl) printing all results, including pseudonyms (dry-run-full-output)
+
+    logprep.pex tests/testdata/config/config-dry-run.yml --dry-run tests/testdata/input_logdata/wineventlog_raw.jsonl --dry-run-input-type jsonl --dry-run-full-output
+"""
 
 import json
 import shutil
 import tempfile
 from copy import deepcopy
 from difflib import ndiff
-from pathlib import Path
 
 from colorama import Fore, Back
 from ruamel.yaml import YAML
 
 from logprep.runner import Runner
+from logprep.util.configuration import Configuration
+from logprep.util.getter import GetterFactory
 from logprep.util.helper import (
     color_print_line,
     recursive_compare,
     remove_file_if_exists,
     color_print_title,
 )
-from logprep.util.json_handling import dump_config_as_file, parse_jsonl, parse_json
+from logprep.util.json_handling import parse_jsonl, parse_json
 
 yaml = YAML(typ="safe", pure=True)
 
@@ -93,49 +132,23 @@ class DryRunner:
     """Used to run pipeline with given events and show changes made by processing."""
 
     def __init__(self, dry_run: str, config_path: str, full_output: bool, use_json: bool, logger):
-        with open(config_path, "r", encoding="utf8") as yaml_file:
-            self._config_yml = yaml.load(yaml_file)
+        self._tmp_path = tempfile.mkdtemp()
+        self.patched_config_path = Configuration.patch_yaml_with_json_connectors(
+            original_config_path=config_path, output_dir=self._tmp_path, input_file_path=dry_run
+        )
         self._full_output = full_output
         self._use_json = use_json
-        self._config_yml["input"] = {
-            "json_input": {
-                "type": "json_input" if use_json else "jsonl_input",
-                "documents_path": dry_run,
-            }
-        }
-        self._config_yml["output"] = {
-            "jsonl_output": {
-                "type": "jsonl_output",
-                "output_file": dry_run,
-            }
-        }
-        self._config_yml["process_count"] = 1
         self._logger = logger
 
     def run(self):
         """Run the dry runner."""
-        tmp_path = tempfile.mkdtemp()
-        config_path = self._patch_config(tmp_path)
-        patched_runner = get_patched_runner(config_path, self._logger)
+        patched_runner = get_patched_runner(self.patched_config_path, self._logger)
         test_output, test_output_custom, test_output_error = get_runner_outputs(patched_runner)
-        input_path = self._config_yml["input"]["json_input"]["documents_path"]
+        config_yaml = GetterFactory.from_string(self.patched_config_path).get_yaml()
+        input_path = config_yaml.get("input", {}).get("patched_input", {}).get("documents_path")
         input_data = parse_json(input_path) if self._use_json else parse_jsonl(input_path)
         self._print_output_results(input_data, test_output, test_output_custom, test_output_error)
-        shutil.rmtree(tmp_path)
-
-    def _patch_config(self, tmp_path):
-        """Generate a config file on disk which contains the output jsonl files in a tmp dir."""
-        tmp_path = Path(tmp_path)
-        output_file = tmp_path / "output.jsonl"
-        output_file_custom = tmp_path / "output_custom.jsonl"
-        output_file_error = tmp_path / "output_errors.jsonl"
-        output_config = self._config_yml["output"]["jsonl_output"]
-        output_config["output_file"] = str(output_file)
-        output_config["output_file_custom"] = str(output_file_custom)
-        output_config["output_file_error"] = str(output_file_error)
-        config_path = str(tmp_path / "generated_config.yml")
-        dump_config_as_file(config_path, self._config_yml)
-        return config_path
+        shutil.rmtree(self._tmp_path)
 
     def _print_output_results(self, input_data, test_output, test_output_custom, test_output_error):
         """Call the print methods that correspond to the output type"""