+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `string_splitter` processor to split strings of variable length into lists
 * Add `ip_informer` processor to enrich events with ip information
 * Allow running the `Pipeline` in python without input/output connectors
+* Add `auto_rule_corpus_tester` to test a whole rule corpus against defined expected outputs.
 
 ## v5.0.1